Extracted subviews to their own file
